created chart component and successfully implemented it in transition page
