+++ src/components/chart/chart.html
@@ -1,0 +1,1 @@
+<div><canvas #ctx id="ctx"></canvas></div>
