+++ src/components/chart/chart.scss
@@ -1,0 +1,3 @@
+chart {
+
+}
