+++ src/pages/transition/transition.html
@@ -15,7 +15,7 @@
 
 
 <ion-content padding>
-  <canvas #canvas id="canvas"></canvas>
+  <chart></chart>
   <ion-list>
     <ion-card *ngFor="let area of areas; let i = index">
       <ion-card-header>
