+++ src/pages/transition/transition.ts
@@ -1,9 +1,6 @@
 import { Component, ViewChild } from '@angular/core';
 import { NavController, NavParams } from 'ionic-angular';
-
-import { Chart } from 'chart.js';
-import 'chartjs-plugin-datalabels';
-
+import { ChartComponent } from '../../components/chart/chart';
 
 /**
  * Generated class for the TransitionPage page.
@@ -18,10 +15,9 @@ import 'chartjs-plugin-datalabels';
 })
 export class TransitionPage {
 
-  @ViewChild('canvas') canvas;
+  @ViewChild(ChartComponent) chartComponent;
 
   areas: Array<any>;
-  chart: any;
 
   constructor(public navCtrl: NavController, public navParams: NavParams) {
     this.areas = [
@@ -110,74 +106,6 @@ export class TransitionPage {
             The "Formatter" property is what allows us to display the data names and values.
             */
 
-  ionViewDidLoad() {
-    this.chart = new Chart(this.canvas.nativeElement, {
-      type: 'polarArea',
-      data: {
-        labels: ["Career", "Finance", "Personal \n Growth", "Health", "Family", "Relationships", "Social life", "Attitude"],
-        datasets: [
-          {
-            backgroundColor: ["rgba(0,0,255, .6)", "rgba(255,0,0, .6)", "rgba(128,0,128, .6)", "rgba(0,128,0, .6)", "rgba(255,165,0, .6)", "rgba(0,128,128, .6)", "rgba(255,0,255, .6)", "rgba(0,255,0, .6)"],
-            borderColor: "black",
-            data: [0, 0, 0, 0, 0, 0, 0, 0]
-          }
-        ]
-      },
-      options: {
-
-        layout: {
-          padding: {
-            top: 55,
-            bottom: 55
-          }
-        },
-
-        legend: {
-          display: false
-        },
-        
-        plugins: {
-          datalabels: {
-            textAlign: 'center',
-            anchor: 'start',
-            align: 'end',
-            offset: function (context) {
-              let chartWidth = context.chart.width;
-              return chartWidth / 4.0 - 60;
-            },
-            backgroundColor: ["rgba(0,0,255, .2)", "rgba(255,0,0, .2)", "rgba(128,0,128, .2)", "rgba(0,128,0, .2)", "rgba(255,165,0, .2)", "rgba(0,128,128, .2)", "rgba(255,0,255, .2)", "rgba(0,255,0, .2)"],
-            borderColor: 'black',
-            rotation: function (context) {
-              if (context.dataIndex === 0 || context.dataIndex === 1 || context.dataIndex === 6 || context.dataIndex === 7) { return 45 / 2 + (45 * context.dataIndex) }
-              else { return 45 / 2 + (45 * context.dataIndex) + 180 }
-            },
-            borderRadius: 5,
-            borderWidth: 0,
-            color: 'black',
-            font: function (context) {
-              var width = context.chart.width;
-              var size = Math.round(width / 60);
-              return {
-                size: size,
-                weight: 600,
-                font: 'Lato'
-              };
-            },
-            formatter: function (value, context) {
-              return context.chart.data.labels[context.dataIndex] + ' ' + context.chart.data.datasets[0].data[context.dataIndex];
-            }
-          }
-        }
-      }
-    });
-
-    //This code down below is to fix a bug where I could not modify the scale in the options
-    //key in the above code (It threw a typescript error even though it worked)
-    this.chart.config.options.scale.ticks.min = 0;
-    this.chart.config.options.scale.ticks.max = 10;
-    this.chart.update();
-  }
-
   toggleSection(area) {
     console.log(area);
     if (area.expand) {
@@ -188,8 +116,8 @@ export class TransitionPage {
   }
 
   changeData(categoryIndex, newNumber) {
-    this.chart.data.datasets[0].data[categoryIndex] = newNumber;
-    this.chart.update();
+    this.chartComponent.chart.data.datasets[0].data[categoryIndex] = newNumber;
+    this.chartComponent.chart.update();
   }
 
 }
